+++ jpro-routing/core-test/src/test/scala/one/jpro/platform/routing/crawl/TestMemoryTester.scala
@@ -19,9 +19,9 @@ class TestMemoryTester {
   @Test
   def simpleTest(): Unit = {
     def route = Route.empty()
-        .and(Route.get("/", r => Response.view(new Page1)))
-        .and(Route.get("/page2", r => Response.view(new Page2)))
-        .and(Route.get("/page4", r => Response.view(new Page2)))
+        .and(Route.get("/", r => Response.page(new Page1)))
+        .and(Route.get("/page2", r => Response.page(new Page2)))
+        .and(Route.get("/page4", r => Response.page(new Page2)))
     val result = AppCrawler.crawlRoute("http://localhost", () => route)
     MemoryTester.testForLeaks(result, () => route)
   }
@@ -30,9 +30,9 @@ class TestMemoryTester {
   def simpleFailingTest(): Unit = {
     val page2 = new Page2
     def route = Route.empty()
-        .and(Route.get("/", r => Response.view(new Page1)))
-        .and(Route.get("/page2", r => Response.view(page2)))
-        .and(Route.get("/page4", r => Response.view(new Page2)))
+        .and(Route.get("/", r => Response.page(new Page1)))
+        .and(Route.get("/page2", r => Response.page(page2)))
+        .and(Route.get("/page4", r => Response.page(new Page2)))
     val result = AppCrawler.crawlRoute("http://localhost", () => route)
     intercept[Throwable](MemoryTester.testForLeaks(result, () => route))
   }
@@ -42,9 +42,9 @@ class TestMemoryTester {
     var node2 = new Label()
 
     def route = Route.empty()
-        .and(Route.get("/", r => Response.view(new Page1)))
+        .and(Route.get("/", r => Response.page(new Page1)))
         .and(Route.get("/page2", r => Response.node(node2)))
-        .and(Route.get("/page4", r => Response.view(new Page2)))
+        .and(Route.get("/page4", r => Response.page(new Page2)))
 
     val result = AppCrawler.crawlRoute("http://localhost", () => route)
     intercept[Throwable](MemoryTester.testForLeaks(result, () => route))
@@ -55,9 +55,9 @@ class TestMemoryTester {
   def simpleFailingTest3(): Unit = {
 
     val route: Route = Route.empty()
-      .and(Route.get("/", r => Response.view(new Page1)))
+      .and(Route.get("/", r => Response.page(new Page1)))
       .and(Route.get("/page2", r => Response.node(new LeakStage)))
-      .and(Route.get("/page4", r => Response.view(new Page2)))
+      .and(Route.get("/page4", r => Response.page(new Page2)))
     val result = AppCrawler.crawlRoute("http://localhost", () => route)
     intercept[Throwable](MemoryTester.testForLeaks(result, () => route)) // fails because the webapp is not collectable
   }