+++ jpro-routing/core/src/main/scala/one/jpro/platform/routing/NodePage.scala
@@ -1,0 +1,7 @@
+package one.jpro.platform.routing
+
+import simplefx.all._
+
+trait NodePage extends Page { this: Node =>
+  final override def content: NodePage with Node = this
+}