+++ jpro-routing/core/src/main/scala/one/jpro/platform/routing/NodeView.scala
@@ -1,7 +1,0 @@
-package one.jpro.platform.routing
-
-import simplefx.all._
-
-trait NodeView extends View { this: Node =>
-  final override def content: NodeView with Node = this
-}